backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
openai>=1.82.0
docling>=2.0.0
pymupdf>=1.25.0
python-dotenv>=1.1.0
Pillow>=11.0.0
